set fixed size on the main window
scalable content or other UI widgets can resize and put part of the
UI out of the screen/display space, this is addressed in agressive
way with this commit - if, for an exmaple, one changes the display
resolution while running the workspace the main window will not be
adjusted. Better solution for the problem will be investigated in
the future.