improvements all around
the SPM triggers support has been rewritten to use configuration
file. It supports a few fields which may change later on but for
now they seems sufficient, maybe different commands for
post-install and post-remove should be used but that is to be
decided.

man pages path fallback has been extended to support paths not only
on /usr but on / too.

the python byte-compiler feature run on every file due to faulty
logic in the site directories checker, that has been corrected.

using short "-f" and "-l" arguments in misc.system_chroot() makes
it Busybox compatibale in addition to the util-linux mount
implementation.

command passed to misc.system_command is split by shlex which
usually does better splitting. this may introduce regression so it
may have to be reverted but subprocess.check_call() does not like
working with strings so it will be further tested and evaluated.