ensure module dependencies are copied (mkinitfs)
originally, the module dependencies were handled with modprobe
invocation. then the algorithm was re-written and internal modules
lookup was introduced but it did no handle module dependencies
with me thinking that it would be enough as usually all modules and
their dependencies are loaded. while this worked fine on existing
installations it did not work for the Live CD/DVD case where only
chosen few are put into the initramfs image and their dependencies
have to be present but (as of now) they are not in the configs.
fixing this by adding modules dependencies resolution algorithm.