implement demotation for building as user other than root
this is the first commit on allowing to build as regular user while
preserving root priviledges in the main process. the actuall
priviledges drop can be done only for sub-processes and is not done
everywhere on purpose (some may be because I have not noticed
thought) and the approach is not 100% perfect in terms of security
but at least it somewhat imroves it and with this feature system
files replacement during installation phase should not be possible
avoiding possible corruptions.

do notice that root group is set on purpose during installation
as it is a good thing for the installed to be owned by the root
group by default avoiding possible security issues. it is also
done to avoid use of `fakeroot'.

this feature is experimental for now and should be handled with
care, I've personally done some manual tests but further tests will
show how it handles in different scenarions with different software