fix various type-checking issues and a long standing download bug
the download bug is with the resume request - when the server does
not support a resume request the download fails and the cache (or
the target sources) had to be deleted before building. that was
very annoying but with this commit the url is checking for
"Accept-Ranges" in the response header and if it is not None (should
be "bytes" from my tests) then additional arguments are passed to
the external utilities to tell them to resume the download. the
internal fetcher still lacks the resume feature so it is not
affected by this but it should check for resume support on the
server side once it does.

there is however slight overheath to this as the url is
queried multiple times:

1. ping - check for 404 and such
2. resume - check if resume is supported
3. fetch - the actual download proccess

and all of this is also done by the external utilities I guess so
it is not very optimal. sadly, the standard Python modules (urllib
and urllib2) do not support cetificate check which is bad in terms
of security. once migrated to Python 3000 the external fetchers
support can be dropped as it does support this feature but that
will not happen anytime soon.