do not return from local_rdepends() when local is ignored target
this can cause lots of dependencies to be ignored as the whole
check is done on per-local targets loop base and returning from
the loop when ignored target is early in the order (e.g. filesystem)
then it will not take into account everything that may depend on
the target but starts with the next alphabet (local_all() returns
sorted result).

in addition, simplified the local metadata gathering algorithm