rework fetcher and friends
the request has to be setuped before opening the remote file
otherwise it gets ignored and the whole files gets downloaded
again and appended to the file corrupting it.

in addition, the new way of handling content-lenght that is 0 is
to block until the server is ready to serve the whole file via
retry without negating iretry - this can take forever!