optimize backing up of files
first of all, instead of comparing the size of the file on system
and the size of the archive member the content is compared for
more accurate desicion making when to backup files. as such a new
method was introduced - misc.archive_content().

second, to reduce the time spend on gathering information about
archive members (size, or content via the new method) all of the
work had to be done in one shot on the same archive object thus
misc.archive_size() and misc.archive_content() accept lists or
tuples (operations on single member are rare and not done in this
codebase). for this to libspm now makes two loops - once to compile
argument that will be passed to misc.archive_content() and second
to actually check whihc files to backup with even more variables
asigning which generally impacts performance but the end results
proves worthy.

this has been on the TODO list for quite some time but I finally
decided to work on it.