make the possible values for special arguments more obvious
the case is that bool type arguments always evaluate to True, even
with action='store' and thus the ast workaround is used. however,
that method does not make proper help output. to fix this and make
the supported values for override arguments right a choice has to
be set explicitly.