split debug info if stripping is enabled
this should probably be an optional feature as those who desire to
strip the debug symbos are likely to not want them at all altough
when they do they will have to recompile. the implementation follows
the instruction for doing so from the following URL:

https://sourceware.org/gdb/onlinedocs/gdb/Separate-Debug-Files.html

the debug symbols can be compressed to save some space but since
debugging procedures and debuggers increase the system load by a
lot I'm afraid that this will slow down the whole process even more
so I've opted to hold on onto this to avoid making debugging on old
machines near impossible. further tests will reveal how and if this
is a good idea but so far from a few manual tests it does not break
anything due to odd situations - I've tested with a few random
targets (namely file, xz and glibc).