mostly databases caching improvements
run cache rebuild checks from local_all() and remote_all() to
ensure the checks are done in all cases because local_search() and
remote_search() are called (on purpose) for all other actions.

also, asign REMOTE_CACHE and LOCAL_CACHE as empty upon
initialization and as an alternative check for self sanity. this
is to make the whole database caching work as expected with two
instances of libpackage.Database() (e.g. different application).
do not that the whole idea of caching does not work with more than
that - every third, fourth, etc. client will face issues. in the
future a database lock to prevent more than two clients should be
implemented.