remove bundled python-dialog
its purpose was as a glue for writing a ncurses frontend for spm,
that is not going to emerge however as I do no feel the need for
one right now. by removing it maintaining it with both Python 2
and Python 3 is no longer a burden and avoids the cluster fuck of
licenses this project has become. libinotify is next on my list
but I'll have to live with it for now.