deprecate database.local_footprint() and database.remote_groups()
in the past where the database was not cached it made sense to have
local_footprint() so avoid overheat (as in to decide what to return
from local_metadata()) but now they are all in the same dictionary
so local_footprint() is all-in-all dublication of code (just
compare it with local_metadata() and see for yourself).

as to remote_groups() - it was inteded for use in the GUI. it never
made it to be utilized properly and it can not be - if groups are
to be used in GUI they must have far more info than just "foo",
for an example sub-group(s), icons, targets in the group, etc.
which is out of the scope of SPM (at least for now). if one wants
the fancy GUI features he can always maintain the data needed in
the repo and write a simple client for that, heck, I would accept
patches too but it has to tought out well.