allow raise of exceptions from libspm by setting CATCH
this has been implemented into libmisc and works well for catching
subprocess exceptions and such from frontends. to further improve
the libraries usability for frontends (e.g. spm-gui) doing the
same for libspm makes a lot of sense. changes are minimal but
there is a heck of difference when something goes wrong and the
frontends is supposed to give feedback to the user.