offload mirrors handling to the fetcher
aside from reducing code dublication by handling mirroring via
arguments instead of ping, check, fetch suquence this will allow
in the future to improve (hopefully) the whole process by relying
on checksums hosted on the mirrors (e.g. foo.sha512 or foo.md5)
or even the original url by extending the fetcher to check for
such file and make use of it. not every source code tarballs host
provides such files but for an example GNOME servers do. with
such functionallity in place the unreliable re-fetch check based
on remote file size will get far more reliable and in addition it
can probably be extended to fetch .sig and verify GPG/PGP
signatures without the need to specify that in the sources array
in SRCBUILDs for an example - blackmagic!