reduce code dublication in terms of offline option handling
misc.fetch() (and friends) and misc.url_ping() already have the
to ensure they do not do any internet activity if misc.OFFLINE is
set to True. calling misc.url_ping() just to print a warning is
not much use and now when internet access is stricktly required
urllib will raise an exception or even libspm itself will do it
(in binary mode).

Git repositories are handled a bit differently because they are not
wrapped around misc.fetch() (but maybe they should) so for them
checking the libspm OFFLINE variable is neccessary.