use custom GnuPG home directory for SPM related features
using the home directory of the user is not desired as SPM will
import keys automatically and if run as the root user (which it
usually is) then that can be (at the very least) a security risk
if you want to later verify some signatures of files not with SPM.

as such libmisc now has a GPGHOME variable that SPM and SPMT
override to /etc/spm/gpg for their own purposes so keys are
imported there. and since the directory must have 0700 permissions
libmisc.Misc.dir_create() now has an additional ipermissions
argument according to which it sets the permissions of the last
directory to be created not its parrents.