make the GnuPG home directory configurable
the reason to make it such is simple - setting it to temporary
directory such as /tmp/spm-gpg one can make the keys imported
temporary not affecting the trusted database at all, or at least
for the next run/reboot (depending on the setup). in such case the
keys will always be imported afresh and the trusted database will
be empty and the risk of someone else importing a key as trusted
into the database will be unlikely, unless he knows the exact path
you set but if you randomize the path (e.g. via mktemp) and set it
via the spm command line argument then the risk is even lower -
perfect for a scripted build-bot and the like.