do not add files to watcher when recursing
to avoid reaching the limit of inotify not adding files to the
watcher when libmisc.INotify.watch_add() is told to recurse is a
decent solution but not perfect.

to see the limit on your system do one of the following:

$ cat /proc/sys/fs/inotify/max_user_watches

on Linux 3.18.9 it is 8192 and that was previously reached fast
with several libpackage.Database clients.