do not store password in plain format
altought encoding the password via base64 is not entirely secure
it is better than nothing. GnuPG uses some sort of read-only memory
storage or something like this buyt I doubt any of that can be
achieved for the purpose of this project since that would require
more low-level programming language. in any case, with this change
the password is less obvious but there should be a way to get it
since it's stored in memory as a known variable with known encoding.

the whole reason not to use pinentry was mentioned as a comment in
the source of libmisc.Misc.getpass() but I'll mention it again -
pinentry does not work well when X (Xorg server) is not run from
the user of SPM (rather libmisc, where the PGP magic happens) and
can block any frontend on top of libspm/libmisc which is not bad.
as such a pseudo password agent was implemented to avoid requiring
the password via user input over and over again. it has its use
case in build-bots too - once can pre-set a base64 encoded password
for the bot in libmisc so that, for an example, building and signing
the tarball happens without interaction.