use ctypes.util.find_library() to find libc needed by libmisc
the C library that GNU C Library (glibc) provides is libc.so.6,
musl provides libc.so and uClibc provides libc.so.0. ctypes
requires the versioned library to be loaded otherwise errors such
as the following one are raised:

/lib/libc.so: invalid ELF header

to avoid this and improve compatibility with other C libraries
the library is looked up via a ctypes method.

this change has not been tested on system with musl/uClibc and is
just theoretically right.