setup the database caches watcher post-initialization
if the watcher is set upon libpackage.Database initializtion and
later the respective ROOT_DIR, CACHE_DIR and LOCAL_DIR are changed
the notified wouldn't respect that. to make it work when the root
directory is changed setup it when the cache is to be created. in
addition add two tests for the caching to ensure it works as
expected.