improve Python 3000 support
the type-checking is just ignored for it, there is no easy way to
support both so addition overheath is in place to do it only if
run from Python version lower than 3. the type-checking tests are
skipped too in such case.

in addition, various string/bytes handling changes have been made
to avoid runtime errors. Python 3 support is, most likely, far from
complete but slowly the codebase is adjusted for it and most of the
code seems to run fine now (tested with SPM).