do not modify the mirrors reference from the fetcher
libmisc.Misc.fetch() removed values from the mirrors list/tuple
passed to it so that it can skip those on the next retry but since
it's a reference it changed the value in other modules namespaces
too which is horrible. to avoid this a local variable is created
that is to be changed.

the issue affected mostly spm's binary mode because it heavily
relies on mirrors for the tarballs because there isn't original
source/url for them.