move the fetcher size recorder where it belongs
libmisc.Misc.fetch_plain() is the only method that uses .last files
to record the remote file sizes so to make the method more robust
the .last file should be created by itself not by other method as
it was previously (libmisc.Misc.fetch()).

with this change the .last file is created on failures too so that
makes the fetcher even better because if an attempt to fetch url
is made, it fails and (say) 10min. later the fetch is retried the
remote file may not be the same. that should now be handled
properly.