make the fetcher re-fetch the signature when URL size changes
basicly this avoids an issue when a file is re-uploaded along with
new signature but there is no difference in the signature sizes
and the fetcher did not force re-fetch the signature. this caused
re-uploaded files with valid signatures to be non-valid and GnuPG
errors out.

now whenever the file changes the local copy of the signature is
removed and re-downloaded to ensure they match.