workaround some lib64 issues on x86_64 hosts
when files get installed in /lib64 or /usr/lib64 and those are
symbolic links (as in LFS/BLFS) the generated libraries cache from
ldconfig does not take that into account and holds the real paths.
this can cause problems because the footprints of targets have
symlinked entries (e.g. /lib64/libfoo.so) but because the cache
does not looking up for the owners of the resolved entries will
say that there is not match.