fixup the shebang detection with space in it
the special \d is for decimal not delimiter (doh) so instead of it
set the shebang detection to look for shebangs with plain spaces
or tabs (\t) in them. in addition, when no match is found instead
of printing a non-sensible path because the match was transformed
into real path add a check for just that case and do not morph it.