improve srcmake binaries detection
making an additional check to see if the binary (e.g. tar) will
work can save the trouble of editing srcmake when it is used to
build software on system which has broken prefered alternative,
for an example "bsdtar" is broken due to openssl library version
bump but "tar" is a symlink to static busybox which will work and
srcmake can and will use now.