make it possible to remove local but not remote target
this was a long-standing issue - if target was removed from the
repositories removing it was impossible without creating a fake
directory with fake SRCBUILD in it or other dirty tricks.

to fix this a copy of the SRCBUILD alone (no sources, patches,
etc. are required) is created in the local directory of the
target at install time from source mode (libspm.Source).

this change will make only work once the targets are rebuild or
SRCBUILDs are manually copied from the repositories to the
respective directory (e.g. copy /var/cache/spm/joe/zlib/SRCBUILD
to /var/local/spm/zlib/SRCBUILD) but once that's done removing
targets will be far more reliable and less tricky (for some rare
cases, as explained above).