initial D-Bus daemon and reworked GUI
the GUI is ment to make use of the daemon but it does not right
now, that will change. there may be some bugs lurking too. in any
case it's step forward towards PackageKit-like implementation
where a daemon gets the job done across GUI sessions along with
some other goodies.