generate recovery initramfs images too
the difference between the regularand the recovery image is that
it includes all modules for the kernel specified (or detected). I
plan on improving that in the future like proving common modules
(e.g. for USB keyboards) automatically from the recovery image
early to improve hardware compatibility in various use cases such
as moving the entire OS, by cloning or whatever, to another
machine that does not have similar hardware and requires different
kernel modules then those that were detected by mkinitfs when
creating the "optimized" image.

sadly, GRUB(2) does not pick the recovery image automatically, so
either it will have to patched or an additional script installed
in /etc/grub.d so beware - manual interaction with bootloaders
configs may be required for this (I have personally tested only
with GRUB 2.02~beta2).