make mkinitfs utilize busybox and handle symlinks better
if the root filesystem (initramfs) overlay had symlinks in it
they could end up messing up the image creation process because
busybox was installed before copying the overlay thus creating
the leading directory which later can not be replaced with
the symlink, this is fixed now.

also, making shutil.copytree() create symlinks avoids some issues
that (again) can be caused from symlinks in the root overlay and
ensures the copy of the original (source) directory is exactly
the same in the initramfs image.

finally, by calling busybox applets instead of external utilities
dependencies for creating the initramfs image are reduced.
depending on the host kmod, findutils and cpio are no longer
required to be in PATH but must be a busybox applet. as a bonus
it saves some lookups in PATH for external utilities which makes
the whole image creation process a bit faster.