implement configurable paths purging
previously, the common conflicts were unconditionally removed.
now this can be changed and in addition it allows removing more
paths (both files and directories) automatically by using a well
formed regular expression.

do note that the override for this feature can only negate it
because it will make the options array more complex if support
in the SRCBUILDs for something like this is implemented (just as
an example):

options=('purge="/usr/share/man"')

so I guess it will stay like this for a while. it may slow down
the indexing process a little bit because now there is no way to
avoid indexing all paths before (now) re-indexing only the files
as the manual pages code also requires it.

and for those interested in the 1.7.x branch - I do not think
I'll be backporting this feature.