speed-up MIME types checking
of the side of libmisc.Misc the optimizations are simple - given
the full path to the file (resolved and otherwise) some common
MIME types can be guessed and avoid using libmisc.Magic. more
types may be added later on to libmisc.Misc.file_mime() later on
for further optimizations.

libmisc.Magic now does not update the access time and does not
check the encoding of the files. I have no idea (yet) how the
access time optimization will work on different filesystem (aside
from the common sense thought that not updating the access time
can reduce the disk I/O), the encodings are not needed in most
cases so skipping some checks on those should speed up the
operations.

the first type of optimizations brings bigger difference when
installing software with lots of files (e.g. icon themes, kernel
sources, etc.) but because they are not reliable they are used
only in the case of SPM's installation procedure right now.

do note that I have not tested the code with Python 3000.