fix some regressions in libspm
libspm.Source.post_update_triggers() was adding leading slash when
joining the list passed which is no longer requires and made some
triggers not execute, that is no longer the case.

some print behavior changes were introduced when the local targets
metadata was changed to JSON format and the some values are now
always saved as list of unicode strings. aside from having the 'u'
prefix they were not printed properly when plain format was
requeste from spm's local/remote mode - the behaviour has been
restored were possible, where it hasn't been is because Python 3
support is in place and it requires unicode data types. when spm
is used on top of Python 3000 all is fine and the 'u' prefix is
not visible on print() and such.