optimize remote targets lookup
since SPM supports both paths and targets from repositories being
passed to it used to check if the target is full path and use the
parser to parse the target's SRCBUILD, the problem was that the
metadata may have been in the cache already if the target is from
the repositories and is full path
(e.g. /var/cache/spm/repositories/mini.git/base/zlib) that
libpackage.Database.remote_search() returns when basenames were
not requested.

by checking if the target is in the cache first and then falling
back to the SRCBUILD parser disk reads can be avoided speeding up
the metadata gathering process.

in addition, for the consistent return values the key a check is
being made wheather the key is in the known array keys instead of
using loop and compare that is (not much) faster.