drop global remote/local cache
it has become harder and harder to keep it working - updating it
(from 0) too often and even forcing it here and there. it may have
made sense before changing the local targets metadata format and
improving the SRCBUILDs parser but now it doesn't and is more
trouble than worth it.