prefix some variables in libmisc.Misc, improve rsync fetcher
there are still a lot of type-unprefixed arguments in libmisc.Misc
but they will probably change in the next minor release.

as for the rsync fetcher - here is the newly used arguments help
in case you do not have rsync installed and you want to know in
what manner that is an improvement:
-c, --checksum        skip based on checksum, not mod-time & size
-H, --hard-links      preserve hard links
-p, --perms           preserve permissions
-E, --executability   preserve the file's executability

Git preserves permissions so that change basicly makes the
fetchers behave consistenly.