implement digest mode into spm-tools
it does nothing special, the cool thing is that the digest file
that is created is portable JSON file which you can carry on a USB
stick or other media and later verify from it.

the bad thing is that it does not record stats (e.g. permissions)
of the target files and no sanity check is done when loading the
digest file during verification which may not have all the fields
required for the verification and will trow error that would
probably not make sense to the end-user.