implement optdepends control
basicly - per-target optdepends preferences, much like package.use
(IIRC). the change set is not small as the GUI has some sort of
support for that too, the daemon method is bogus tough. right now,
the changes are not very strict - unless dependencies are build
(including optdepends via -O) the option will not be enabled as
libpackage.Database.local_uptodate() is used in the check for this
from libspm.Source and the dependency has to be met for it to
export ENABLE_<FEATURE>=yes. this will likely change as I do not
presume that someone request libass support for ffmpeg and does
not want it, as in it has to be enabled no matter what or error
should occur.