implement experimental aport mode into spm
that's basicly a toy, if all goes well (and I bet it will not) it
will be something fancy - you pass it a bunch of URLs to source
code and it should create (if possible) SRCBUILD for each
optionally building it right after that.

it is not very flexible and lacks a few things (like automatic
maintainer assignment), not document too but it was written very
quick for the sake of an experiment. I would like to make it a bit
more robust but it may be dropped without warning.