optimize target files removal on upgrade
by moving out the desicion maker for which files to not remove
inside the target files (not dirs or links) removal loop a big
time spend on looping over all files and creating a new list with
content to be removed is avoided.

it is not extremly fast now either but at least it is faster, as
you can notice the difference set had to be converted to tuple
for it to be reversed for the sake of directories and links
removal but overall this proves to be faster anyway. and since
a set() does not allow being reversed directly this can no be
optimized further without some other drastic code changes in the
procedure, prove me wrong and I'll be happy to merge patches :).