create pid files for ahws and spmd
instead of using signal catching via the signal module I opted to
use the simple try/finally technique for now, with this change
start-stop-daemon can properly start/stop them but the cleanup is
not perfect, for an example if the daemons are killed with
`kill -9 spmd ahws' the pid files will remain until the next
reboot (if /var/run is tmpfs that is). On top of that the pid
files are overwriteen by the daemon which means that multiple
instances can be spawned but using signals catching turned more
tricky than expected - SIGKILL seems to trow invalid argument and
some signals were not caught at all (bugger). it may be just me
not understanding how to use the module but that will do for now.