introduce new local metadata field - autodepends
also, start using it for reverse dependencies checking and add
migrate procedure for it to upgrade mode. the regex lookup used
in the reverse targets check during merge is still slow but
overall it is faster.

check mode has been rewritten, much of its purpose ripped off but
I rarely found use for it and it has been maintenance burden for
quite some time. instead of the deep dependencies scanning it is
using autodepends metadata now. oh, well.

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>