make the optdepends feature not controllable by user
optdepends were supposed to be automatically pulled on need,
supporting argument was consfusing and messy thus dropping it (in
a future micro/patch release, but better now whlile the 1.9.x
series are still hot). this fixes an issue where optdepends were
not build even when outdated and building a target that depends on
that optional dependency.

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>