prepare Copyright for the year to come
I do not think I will be doing anything on the project for the months to
come, unless I _really_ have to. The project also pre-dates 2013, despite
first public release on GitHub being made on Apr 14 2013 the early
prototypes were long before that. The first public commit on my spm repo is
https://github.com/fluxer/spm/commit/602e74aa5dce36e0e5d1b978e9c53cffddcc5e31

Signed-off-by: Ivailo Monev <xakepa10@laimg.moc>